Improve typescript / javascript linting
Automating fixing thanks to eslint

+++ update.sh
@@ -6,7 +6,8 @@ source $DOTFILES/install_config
 
 # Save zsh history file
 
-rm "$DOTFILES_CLOUD/zsh/.zhistory"
+[ -f "$DOTFILES_CLOUD/zsh/.zhistory" ] && rm "$DOTFILES_CLOUD/zsh/.zhistory"
+mkdir -p "$DOTFILES_CLOUD/zsh/"
 cp "$ZDOTDIR/.zhistory" "$DOTFILES_CLOUD/zsh/.zhistory"
 
 #---------------------------------------
